minor-fixes fix: refactor login service
