--- app/Repositories/Contracts/UserRepositoryInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace App\Repositories\Contracts;
-
-interface UserRepositoryInterface
-{
-    public function login(array $credentials): bool;
-    public function logout(): bool;
-}
